FrontEnd/src/app/login/login.component.spec.ts
import { AuthenticationService } from './../services/authentication.service';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { LoginComponent } from './login.component';
import { RouterTestingModule } from '@angular/router/testing';
import { HttpClientModule } from '@angular/common/http';
import { FormsModule, NgForm } from '@angular/forms';
import { Router } from '@angular/router';


describe('LoginComponent', () => {
  let component: LoginComponent;
  let fixture: ComponentFixture<LoginComponent>;
  let router: Router;
  
  

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [ RouterTestingModule,HttpClientModule,FormsModule],
   
      declarations: [ LoginComponent ],
 
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(LoginComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
    router = TestBed.get(Router);
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  
  it('Breadcrum routing', () => {
  
    const accountForm = <NgForm>{

      value :{
        email:'',
        password: ''
      }
    }
    component.login(accountForm);
   
      const navigateSpy = spyOn(router, 'navigate');
      component.login(accountForm);
    if(accountForm.valid)
    { expect(navigateSpy).toBeTruthy();
           } 

           else{
expect(navigateSpy).toBeFalsy();

           }
    
    
  });


});
